Strip build-time behaviours
The import of CSS has temporarily been removed since the next step will
change the addon directory into a proper v2 addon. A migration path (or
solution) will be provided after the V2 scaffhold is in place.

This commit covers part 3 of the Embroider V2 addon conversion guide
See: https://github.com/embroider-build/embroider/blob/main/docs/porting-addons-to-v2.md
