+++ addon/app/initializers/load-bootstrap-config.js
@@ -1,13 +1,1 @@
-import ENV from '../config/environment';
-import Config from 'ember-bootstrap/config';
-import { registerLibrary } from 'ember-bootstrap/version';
-
-export function initialize(/* container, application */) {
-  Config.load(ENV['ember-bootstrap'] || {});
-  registerLibrary();
-}
-
-export default {
-  name: 'load-bootstrap-config',
-  initialize,
-};
+export { default } from 'ember-bootstrap/initializers/load-bootstrap-config';
