+++ docs/gulpfile.js
@@ -26,7 +26,9 @@ gulp.task(
     return merge(
       gulp.src('docs/api/**/*', { base: 'docs' }),
       gulp.src('docs/dist/**/*'),
-      gulp.src('CHANGELOG.md').pipe(transform(striptags, { encoding: 'utf8' })),
+      gulp
+        .src('../CHANGELOG.md')
+        .pipe(transform(striptags, { encoding: 'utf8' })),
     ).pipe(gulp.dest('docs/build'));
   }),
 );
